Game logic/styling
Completed the game logic. Each choice is handled in the api and ui. Game over is checked with every move and if player won, no more spaces can be clicked until new game is started. New game button resets board. Added button to get all previous games and display them as a in carousel in modal. User can scroll through previous games and see exactly how it ended with dummy board for visual. Added delete button to each previous game if user desires to delete a certain past game. Fixed x and o on the board to be centered.